**FJCS** Final version
