+++ BitsoChallenge/bitso-challenge/src/App.js
@@ -5,7 +5,6 @@ import Header from './Components/Header/Header';
 import BodySection from './Components/BodySection/BodySection';
 import {BOOKS_AVAILABLES} from './Config'
 import {URL_SERVICES} from './Config.js';
-import {PARAMS_SERVICES} from './Config.js';
 import {ARRAY_TRADE_DATES} from './Config.js';
 import {getData} from './Utils/CallServices.js';
 import {callGetServices} from './Utils/CallServices.js';
@@ -20,27 +19,31 @@ class App extends Component {
       ticketInfo:null,
       array_trades:null,
       array_post:null,
-      orderBooks:null
+      orderBooks:null,
+      arrayBooks:null
     }
 
+    //Timmer to update info
     this.interval = setInterval(() => {
       this.getAllServices();
   }, 4000)
 
+  this.getAllServices();
     this.changeBook=this.changeBook.bind(this);
     this.changeRangeGraphic=this.changeRangeGraphic.bind(this)
   }
 
-
+  //Call all used services
   getAllServices(){
     this.getBtcCost();
     this.getTrades();
     this.getGraphic();
     this.getOrderBook();
+    this.getBooks();
   }
 
+  //Update info with the boook selected
   changeBook(book){
-    
     this.state.activeCoin=book;
     this.getBtcCost();
     this.getTrades();
@@ -48,12 +51,13 @@ class App extends Component {
     this.getOrderBook();
   }
 
+  //Update date range to chart
   changeRangeGraphic(range){
     this.state.activeTimeTrade=range;
     this.getGraphic();
   }
 
-
+  //Service used in postures
   getOrderBook(){
     var arrayParams=[];
     var objParam={param:'', value:''}
@@ -63,7 +67,6 @@ class App extends Component {
     callGetServices(URL_SERVICES.OrderBook,arrayParams).then(response => {
       if(response!=='error'){
       var sumBid=0;
-      var totAmountBid=0;
       response.payload.bids.forEach(bid => {
         sumBid=(+sumBid)+(+bid.amount);
         bid.valor=(bid.amount*bid.price)
@@ -80,7 +83,6 @@ class App extends Component {
       response.payload.asks.forEach(ask => {
         sumAsk=(+sumAsk)+(+ask.amount);
         ask.valor=(ask.amount*ask.price)
-
         ask.valor=(+ask.valor).toFixed(8);
         ask.amount=(+ask.amount).toFixed(8);
         ask.price=(+ask.price).toFixed(8);
@@ -94,6 +96,7 @@ class App extends Component {
   );
   }
 
+  //Get actual value to initial selected coin
   getBtcCost(){
     var arrayParams=[];
     var objParam={param:'', value:''}
@@ -107,6 +110,7 @@ class App extends Component {
   );
   }
 
+  //Get info used in chart
   getGraphic(){
     getData(this.state.activeCoin.book,this.state.activeTimeTrade.value,URL_SERVICES.TradeP).then(data => {
       if(data!=='error'){
@@ -118,10 +122,10 @@ class App extends Component {
 		})
   }
 
+  //Get trades
   getTrades(){
     var arrayParams=[];
     var objParam={param:'', value:''}
-
     objParam.param='book';
     objParam.value=this.state.activeCoin.book;
     arrayParams.push(objParam)
@@ -132,18 +136,36 @@ class App extends Component {
     }
   );
   }
+  //Get all info books
+  getBooks(){
+    var arrayBooks=[];
+    var arrayParams=[];
+    var objParam={param:'', value:''}
+    objParam.param='book';
+    BOOKS_AVAILABLES.forEach(book => {
+        objParam.value=book.book;
+        arrayParams[0]=objParam;
+        callGetServices(URL_SERVICES.Ticker,arrayParams).then(response => {
+            response.book=book;
+          arrayBooks.push(response);
+          this.setState({arrayBooks})
+    });    
+    }
+
+);  
+}
 
   render() {
-    if(this.state.ticketInfo!==null && this.state.array_trades!==null && this.state.array_post!==null && this.state.orderBooks!==null){
-
-      if(this.state.ticketInfo!=='error' && this.state.array_trades!=='error' && this.state.array_post!=='error' && this.state.orderBooks!=='error'){
+    if(this.state.ticketInfo!==null && this.state.array_trades!==null && this.state.array_post!==null && this.state.orderBooks!==null && this.state.arrayBooks!==null){
+      if(this.state.ticketInfo!=='error' && this.state.array_trades!=='error' && this.state.array_post!=='error' && this.state.orderBooks!=='error' &&this.state.arrayBooks!=='error'){
         return (
           <div className="App">
-           <Header changeBook={this.changeBook} ticketInfo={this.state.ticketInfo} activeCoin={this.state.activeCoin}/>
+           <Header arrayBooks={this.state.arrayBooks} changeBook={this.changeBook} ticketInfo={this.state.ticketInfo} activeCoin={this.state.activeCoin}/>
            <BodySection orderBooks={this.state.orderBooks} changeRangeGraphic={this.changeRangeGraphic} 
            activeTimeTrade={this.state.activeTimeTrade} array_post={this.state.array_post} 
            array_trades={this.state.array_trades} activeCoin={this.state.activeCoin}
-           ticketInfo={this.state.ticketInfo} />
+           ticketInfo={this.state.ticketInfo} 
+           arrayBooks={this.state.arrayBooks} />
           </div>
         );
       }else{
