+++ BitsoChallenge/bitso-challenge/src/Components/BodySection/Markets/Markets.js
@@ -2,9 +2,7 @@ import React, { Component } from 'react';
 import arrow from './../../../Assets/Images/SVG/icon_dropdown.svg';
 import arrowGreen from './../../../Assets/Images/SVG/order_selector_green.svg';
 import arrowRed from './../../../Assets/Images/SVG/order_selector_red.svg';
-import callGetServices from './../../../Utils/CallServices';
-import {URL_SERVICES} from './../../../Config';
-import {BOOKS_AVAILABLES} from './../../../Config';
+
 import './Markets.css';
 
 class Markets extends Component {
@@ -12,7 +10,6 @@ class Markets extends Component {
         super(props);
         
         this.state={
-          arrayBooks:[],
           markets:"",
           btc_to_mxn:0,
           activeCoin:this.props.activeCoin,
@@ -20,51 +17,20 @@ class Markets extends Component {
           array_post:this.props.array_post,
           activeTimeTrade:this.props.activeTimeTrade,
           orderBooks:this.props.orderBooks,
-          ticketInfo:this.props.ticketInfo
+          ticketInfo:this.props.ticketInfo,
+          arrayBooks:this.props.arrayBooks
 
         }
-        this.getBooks();
-        this.interval = setInterval(() => {
-            this.getBooks();
-        }, 4000)
 
-
-        // this.changeRangeGraphic=this.changeRangeGraphic.bind(this)
       }
 
       componentWillReceiveProps(nextProps){
-          
-        // this.setState({activeCoin:nextProps.activeCoin, array_trades:nextProps.array_trades,
-        //   array_post:nextProps.array_post ,activeTimeTrade:nextProps.activeTimeTrade,orderBooks:nextProps.orderBooks,
-        //   ticketInfo:nextProps.ticketInfo})
+        this.setState({arrayBooks:nextProps.arrayBooks})
       }
 
 
 
-      getBooks(){
-            var arrayBooks=[];
 
-            var arrayParams=[];
-            var objParam={param:'', value:''}
-
-
-            objParam.param='book';
-
-            BOOKS_AVAILABLES.forEach(book => {
-                objParam.value=book.book;
-                arrayParams[0]=objParam;
-                callGetServices(URL_SERVICES.Ticker,arrayParams).then(response => {
-                    response.book=book;
-                  arrayBooks.push(response);
-
-                  this.setState({arrayBooks})
-            });    
-            }
-        
-        );
-
-            
-    }
 
 
       changeRangeGraphic(range){
@@ -86,7 +52,8 @@ class Markets extends Component {
                 <p className='columnBook'>{book.book.fromCoin}/{book.book.toCoin}</p>
                 
                 <section className={(book.payload.last<book.payload.vwap?'columnPriceDown':'columnPriceUp')}> <img src={(book.payload.last<book.payload.vwap?arrowRed:arrowGreen)}/><p > {book.payload.last} {book.book.toCoin}</p></section>
-              
+
+
                 </section>
 
             })}
